:recycle: 문제 만들기 - 네이밍 수정 및 폴더링 (#14, #27)
